+++ src/core/compression/compression_func.cpp
@@ -68,7 +68,7 @@ static ERR compress_folder(extCompression *Self, std::string Location, std::stri
    if (!Path.empty()) {
       // Seek to the position at which this new directory entry will be added
 
-      ULONG dataoffset = 0;
+      uint32_t dataoffset = 0;
       if (!Self->Files.empty()) {
          auto &last = Self->Files.back();
          if (acSeekStart(Self->FileIO, last.Offset + HEAD_NAMELEN) != ERR::Okay) {
@@ -113,10 +113,10 @@ static ERR compress_folder(extCompression *Self, std::string Location, std::stri
       copymem(glHeader, header, sizeof(glHeader));
 
       wrb<uint16_t>(entry.DeflateMethod, header + HEAD_DEFLATEMETHOD);
-      wrb<ULONG>(entry.TimeStamp, header + HEAD_TIMESTAMP);
-      wrb<ULONG>(entry.CRC, header + HEAD_CRC);
-      wrb<ULONG>(entry.CompressedSize, header + HEAD_COMPRESSEDSIZE);
-      wrb<ULONG>(entry.OriginalSize, header + HEAD_FILESIZE);
+      wrb<uint32_t>(entry.TimeStamp, header + HEAD_TIMESTAMP);
+      wrb<uint32_t>(entry.CRC, header + HEAD_CRC);
+      wrb<uint32_t>(entry.CompressedSize, header + HEAD_COMPRESSEDSIZE);
+      wrb<uint32_t>(entry.OriginalSize, header + HEAD_FILESIZE);
       wrb<uint16_t>(entry.Name.size(), header + HEAD_NAMELEN);
       if (acWriteResult(Self->FileIO, header, HEAD_LENGTH) != HEAD_LENGTH) return ERR::Okay;
       if (acWriteResult(Self->FileIO, entry.Name.c_str(), entry.Name.size()) != (LONG)entry.Name.size()) return ERR::Okay;
@@ -163,7 +163,7 @@ static ERR compress_file(extCompression *Self, std::string Location, std::string
 
    CSTRING symlink = nullptr;
    bool deflateend = false;
-   ULONG dataoffset = 0;
+   uint32_t dataoffset = 0;
    std::string filename;
    std::list<ZipFile>::iterator file_index;
    LONG i, len;
@@ -387,10 +387,10 @@ static ERR compress_file(extCompression *Self, std::string Location, std::string
    UBYTE header[sizeof(glHeader)];
    copymem(glHeader, header, sizeof(glHeader));
    wrb<uint16_t>(entry.DeflateMethod, header + HEAD_DEFLATEMETHOD);
-   wrb<ULONG>(entry.TimeStamp, header + HEAD_TIMESTAMP);
-   wrb<ULONG>(entry.CRC, header + HEAD_CRC);
-   wrb<ULONG>(entry.CompressedSize, header + HEAD_COMPRESSEDSIZE);
-   wrb<ULONG>(entry.OriginalSize, header + HEAD_FILESIZE);
+   wrb<uint32_t>(entry.TimeStamp, header + HEAD_TIMESTAMP);
+   wrb<uint32_t>(entry.CRC, header + HEAD_CRC);
+   wrb<uint32_t>(entry.CompressedSize, header + HEAD_COMPRESSEDSIZE);
+   wrb<uint32_t>(entry.OriginalSize, header + HEAD_FILESIZE);
    wrb<uint16_t>(entry.Name.size(), header + HEAD_NAMELEN);
    if (acWriteResult(Self->FileIO, header, HEAD_LENGTH) != HEAD_LENGTH) return ERR::Write;
    if (acWriteResult(Self->FileIO, entry.Name.c_str(), entry.Name.size()) != (LONG)entry.Name.size()) return ERR::Write;
@@ -471,7 +471,7 @@ static ERR fast_scan_zip(extCompression *Self)
    if (acSeek(Self->FileIO, TAIL_LENGTH, SEEK::END) != ERR::Okay) return ERR::Seek; // Surface error, fail
    if (acRead(Self->FileIO, &tail, TAIL_LENGTH, nullptr) != ERR::Okay) return ERR::Read; // Surface error, fail
 
-   if (0x06054b50 != ((ULONG *)&tail)[0]) {
+   if (0x06054b50 != ((uint32_t *)&tail)[0]) {
       // Tail not available, use the slow scanner instead
       return scan_zip(Self);
    }
@@ -494,7 +494,7 @@ static ERR fast_scan_zip(extCompression *Self)
       }
 
       #pragma GCC diagnostic ignored "-Waddress-of-packed-member"
-      auto head = (ULONG *)list;
+      auto head = (uint32_t *)list;
       #pragma GCC diagnostic warning "-Waddress-of-packed-member"
 
       for (LONG i=0; i < tail.filecount; i++) {
@@ -553,7 +553,7 @@ static ERR fast_scan_zip(extCompression *Self)
 
          Self->Files.push_back(zf);
 
-         head = (ULONG *)(((UBYTE *)head) + LIST_LENGTH + scan->commentlen + scan->namelen + scan->extralen);
+         head = (uint32_t *)(((UBYTE *)head) + LIST_LENGTH + scan->commentlen + scan->namelen + scan->extralen);
       }
    }
 
@@ -581,8 +581,8 @@ static ERR scan_zip(extCompression *Self)
          if (acSeek(Self->FileIO, (double)HEAD_COMPRESSEDSIZE-4, SEEK::CURRENT) != ERR::Okay) return log.warning(ERR::Seek);
 
          struct {
-            ULONG compressedsize;
-            ULONG originalsize;
+            uint32_t compressedsize;
+            uint32_t originalsize;
             uint16_t namelen;
             uint16_t extralen;
          } header;
@@ -719,28 +719,28 @@ static void write_eof(extCompression *Self)
          if (fl::ReadLE(Self->FileIO, &namelen) != ERR::Okay) return;
          if (fl::ReadLE(Self->FileIO, &extralen) != ERR::Okay) return;
          acSeekCurrent(Self->FileIO, last.CompressedSize + namelen + extralen);
-         ULONG listoffset = last.Offset + last.CompressedSize + namelen + extralen + HEAD_LENGTH;
+         uint32_t listoffset = last.Offset + last.CompressedSize + namelen + extralen + HEAD_LENGTH;
 
          // Write out the central directory
 
-         ULONG listsize  = 0;
+         uint32_t listsize  = 0;
          uint16_t filecount = 0;
          for (auto &chain : Self->Files) {
             UBYTE elist[sizeof(glList)];
             copymem(glList, elist, sizeof(glList));
 
             wrb<uint16_t>(chain.DeflateMethod, elist+LIST_METHOD);
-            wrb<ULONG>(chain.TimeStamp, elist+LIST_TIMESTAMP);
-            wrb<ULONG>(chain.CRC, elist+LIST_CRC);
-            wrb<ULONG>(chain.CompressedSize, elist+LIST_COMPRESSEDSIZE);
-            wrb<ULONG>(chain.OriginalSize, elist+LIST_FILESIZE);
+            wrb<uint32_t>(chain.TimeStamp, elist+LIST_TIMESTAMP);
+            wrb<uint32_t>(chain.CRC, elist+LIST_CRC);
+            wrb<uint32_t>(chain.CompressedSize, elist+LIST_COMPRESSEDSIZE);
+            wrb<uint32_t>(chain.OriginalSize, elist+LIST_FILESIZE);
             wrb<uint16_t>(chain.Name.size(), elist+LIST_NAMELEN);
             wrb<uint16_t>(0, elist+LIST_EXTRALEN);
             wrb<uint16_t>(chain.Comment.size(), elist+LIST_COMMENTLEN);
             wrb<uint16_t>(0, elist+LIST_DISKNO);
             wrb<uint16_t>(0, elist+LIST_IFILE);
-            wrb<ULONG>(chain.Flags, elist+LIST_ATTRIB);
-            wrb<ULONG>(chain.Offset, elist+LIST_OFFSET);
+            wrb<uint32_t>(chain.Flags, elist+LIST_ATTRIB);
+            wrb<uint32_t>(chain.Offset, elist+LIST_OFFSET);
 
             acWriteResult(Self->FileIO, elist, LIST_LENGTH);
 
@@ -756,8 +756,8 @@ static void write_eof(extCompression *Self)
 
          wrb<uint16_t>(filecount,  tail + TAIL_FILECOUNT); // File count for this file
          wrb<uint16_t>(filecount,  tail + TAIL_TOTALFILECOUNT); // File count for all zip files when spanning multiple archives
-         wrb<ULONG>(listsize,   tail + TAIL_FILELISTSIZE);
-         wrb<ULONG>(listoffset, tail + TAIL_FILELISTOFFSET);
+         wrb<uint32_t>(listsize,   tail + TAIL_FILELISTSIZE);
+         wrb<uint32_t>(listoffset, tail + TAIL_FILELISTOFFSET);
          acWriteResult(Self->FileIO, tail, TAIL_LENGTH);
       }
       else Self->FileIO->set(FID_Size, 0);
